requirements.txt
appdirs==1.4.4
asgiref==3.3.4
backports.entry-points-selectable==1.1.0
black==21.6b0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
Django==3.2.4
django-cors-headers==3.7.0
django-import-export==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
et-xmlfile==1.1.0
filelock==3.0.12
idna==2.10
isort==5.9.2
itypes==1.2.0
Jinja2==3.0.1
MarkupPy==1.14
MarkupSafe==2.0.1
mypy-extensions==0.4.3
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.7
pathspec==0.8.1
Pillow==8.3.1
pipenv==2021.5.29
platformdirs==2.0.2
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.5
virtualenv==20.6.0
virtualenv-clone==0.5.4
xlrd==2.0.1
xlwt==1.3.0
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-ckeditor==6.0.0

flower==1.0.0
celery==5.1.2
redis==3.5.3

drf-yasg==1.20.0